requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.7
werkzeug>=3.0.0

pandas>=2.3.0
openpyxl==3.1.2
gigachat>=0.2.0
python-dotenv==1.0.0
numpy>=2.2.0
urllib3>=1.26.0
pdfplumber>=0.11.9
python-docx>=1.1.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
